Create transaction.rs
