+++ banks-client/src/error.rs
@@ -1,5 +1,5 @@
 use {
-    solana_transaction_context::TransactionReturnData,
+    solana_transaction_context::transaction::TransactionReturnData,
     solana_transaction_error::{TransactionError, TransportError},
     std::io,
     tarpc::client::RpcError,
