+++ cli-output/src/display.rs
@@ -737,7 +737,7 @@ mod test {
         solana_pubkey::Pubkey,
         solana_signer::Signer,
         solana_transaction::Transaction,
-        solana_transaction_context::TransactionReturnData,
+        solana_transaction_context::transaction::TransactionReturnData,
         solana_transaction_status::{Reward, RewardType, TransactionStatusMeta},
         std::io::BufWriter,
     };
