+++ program-runtime/src/lib.rs
@@ -24,6 +24,6 @@ pub mod __private {
         solana_hash::Hash,
         solana_instruction::error::InstructionError,
         solana_rent::Rent,
-        solana_transaction_context::TransactionContext,
+        solana_transaction_context::transaction::TransactionContext,
     };
 }
