+++ program-test/tests/return_data.rs
@@ -11,7 +11,7 @@ use {
     solana_pubkey::Pubkey,
     solana_signer::Signer,
     solana_transaction::Transaction,
-    solana_transaction_context::TransactionReturnData,
+    solana_transaction_context::transaction::TransactionReturnData,
     std::{slice, str::from_utf8},
 };
