+++ storage-proto/src/lib.rs
@@ -7,7 +7,7 @@ use {
     },
     solana_message::v0::LoadedAddresses,
     solana_serde::default_on_eof,
-    solana_transaction_context::TransactionReturnData,
+    solana_transaction_context::transaction::TransactionReturnData,
     solana_transaction_error::{TransactionError, TransactionResult as Result},
     solana_transaction_status::{
         InnerInstructions, Reward, RewardType, TransactionStatusMeta, TransactionTokenBalance,
