+++ svm-test-harness/instr/src/harness.rs
@@ -18,7 +18,7 @@ use {
     solana_svm_log_collector::LogCollector,
     solana_svm_timings::ExecuteTimings,
     solana_svm_transaction::{instruction::SVMInstruction, svm_message::SVMStaticMessage},
-    solana_transaction_context::TransactionContext,
+    solana_transaction_context::transaction::TransactionContext,
     std::{collections::HashSet, rc::Rc, sync::Arc},
 };
