+++ svm/tests/integration_test.rs
@@ -52,7 +52,7 @@ use {
     solana_system_transaction as system_transaction,
     solana_sysvar::rent::Rent,
     solana_transaction::{Transaction, sanitized::SanitizedTransaction},
-    solana_transaction_context::TransactionReturnData,
+    solana_transaction_context::transaction::TransactionReturnData,
     solana_transaction_error::TransactionError,
     std::{collections::HashMap, num::NonZeroU32, slice, sync::atomic::Ordering},
     test_case::test_case,
